requirements.txt
streamlit
pandas
openai
anthropic
python-dotenv
pdfplumber
langchain
unstructured
plotly
